+++ e2e/wallets/bip85-subaccounts/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @coinmasters/integration-coins
 
+## 1.12.1
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.1
+  - @coinmasters/types@4.10.2
+
+## 1.12.0
+
+### Minor Changes
+
+- bump
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.0
+
+## 1.11.19
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.12.22
+
+## 1.11.18
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.12.21
+
 ## 1.11.17
 
 ### Patch Changes