+++ e2e/wallets/intergration-kkapi/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @coinmasters/integration-kkapi
 
+## 1.0.8
+
+### Patch Changes
+
+- bump
+
 ## 1.0.7
 
 ### Patch Changes